+++ SkyFeedConnectAPI/src/main/java/com/Ada/SkyFeedConnect/repository/UserRepository.java
@@ -5,6 +5,17 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
+
+/**
+ * Repository interface for User entities.
+ */
 public interface UserRepository extends JpaRepository<User, Long> {
-  Optional<User> findByUsername(String username);
+
+    /**
+     * Finds a user by username.
+     *
+     * @param username The username to search for.
+     * @return An Optional containing the user if found, otherwise empty.
+     */
+    Optional<User> findByUsername(String username);
 }
