+++ vite-project/src/components/features/NewsFeed/NewsFeed.jsx
@@ -28,15 +28,19 @@ const NewsFeed = () => {
       <div className="main-news-item">
         {newsItems.length > 0 && (
           <>
-            <h1 className="main-news-subtitle"><a href = {newsItems[0].link}>{newsItems[0].titulo}</a></h1>
-            <a href = {newsItems[0].link}>
-            <img
-              src={newsItems[0].imagens}
-              alt={newsItems[0].titulo}
-              className="secondary-news-image"
-            />
+            <h1 className="main-news-subtitle">
+              <a href={newsItems[0].link}>{newsItems[0].titulo}</a>
+            </h1>
+            <a href={newsItems[0].link}>
+              <img
+                src={newsItems[0].imagens}
+                alt={newsItems[0].titulo}
+                className="secondary-news-image"
+              />
             </a>
-            <p className="main-news-summary">{newsItems[0].introducao}</p>
+            <p className="main-news-summary">
+              <a href={newsItems[0].link}>{newsItems[0].introducao}</a>{" "}
+            </p>
           </>
         )}
       </div>
@@ -45,16 +49,24 @@ const NewsFeed = () => {
         {newsItems.slice(1, 4).map((news) => (
           <div key={news.id} className="secondary-news-item">
             {news.imagens && (
-              <img
-                src={news.imagens}
-                alt={news.titulo}
-                className="secondary-news-image"
-              />
+              <a href={news.link}>
+                <img
+                  src={news.imagens}
+                  alt={news.titulo}
+                  className="secondary-news-image"
+                />
+              </a>
             )}
             <div className="secondary-news-text">
-              <h2 className="secondary-news-title">{news.editorias} </h2>
-              <h3 className="secondary-news-subtitle"><a href={news.link}>{news.titulo}</a></h3>
-              <p className="secondary-news-summary">{news.introducao}</p>
+              <h2 className="secondary-news-title">
+                <a href={news.link}>{news.editorias}</a>{" "}
+              </h2>
+              <h3 className="secondary-news-subtitle">
+                <a href={news.link}>{news.titulo}</a>
+              </h3>
+              <p className="secondary-news-summary">
+                <a href={news.link}>{news.introducao}</a>{" "}
+              </p>
             </div>
           </div>
         ))}
@@ -64,9 +76,15 @@ const NewsFeed = () => {
         {newsItems.slice(4).map((news) => (
           <div key={news.id} className="additional-news-item">
             <div className="additional-news-text">
-              <h2 className="additional-news-title">{news.editorias}</h2>
-              <h3 className="additional-news-subtitle"><a href = {news.link}>{news.titulo}</a></h3>
-              <p className="additional-news-summary">{news.introducao}</p>
+              <h2 className="additional-news-title">
+                <a href={news.link}>{news.editorias}</a>
+              </h2>
+              <h3 className="additional-news-subtitle">
+                <a href={news.link}>{news.titulo}</a>
+              </h3>
+              <p className="additional-news-summary">
+                <a href={news.link}>{news.introducao}</a>
+              </p>
             </div>
           </div>
         ))}
